project/requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
gunicorn==20.1.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyserial==3.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
flask-sqlalchemy==2.5.1
